Refs 7289 - Changes made as per the peer reviewer comments
