+++ app/models/organization.rb
@@ -63,7 +63,7 @@ class Organization
   embeds_one :general_agency_profile, cascade_callbacks: true, validate: true
   embeds_one :carrier_profile, cascade_callbacks: true, validate: true
   embeds_one :hbx_profile, cascade_callbacks: true, validate: true
-  embeds_many :invoices, as: :documentable
+  embeds_many :documents, as: :documentable
 
   accepts_nested_attributes_for :office_locations, :employer_profile, :broker_agency_profile, :carrier_profile, :hbx_profile, :general_agency_profile
 
@@ -75,6 +75,8 @@ class Organization
     uniqueness: true
 
   validate :office_location_kinds
+
+  BUCKET_NAME = "invoices"
 
   index({ hbx_id: 1 }, { unique: true })
   index({ legal_name: 1 })
@@ -142,6 +144,10 @@ class Organization
     write_attribute(:hbx_id, HbxIdGenerator.generate_organization_id) if hbx_id.blank?
   end
 
+  def invoices 
+    documents.select{ |document| document.subject == 'invoice' }
+  end
+
   # Strip non-numeric characters
   def fein=(new_fein)
     write_attribute(:fein, new_fein.to_s.gsub(/\D/, ''))
@@ -202,6 +208,55 @@ class Organization
 
   def self.valid_carrier_names_for_options
     Organization.valid_carrier_names.invert.to_a
+  end
+
+  def self.upload_invoice(file_path)
+    invoice_date = get_invoice_date(file_path) rescue nil
+    org = get_organization(file_path) rescue nil
+    if invoice_date && org && !invoice_exist?(invoice_date,org)
+      s3file= Aws::S3Storage.save(file_path, BUCKET_NAME)
+      document = Document.new
+      if s3file
+        document.identifier = s3file
+        document.type ="pdf"
+        document.date = invoice_date
+        document.format = 'application/pdf'
+        document.subject = 'invoice'
+        document.title = get_file_name(file_path)
+        org.documents << document
+        logger.debug "associated file #{file_path} with the Organization"
+        return document
+      end
+    else
+      logger.warn("Unable to associate invoice #{file_path}")
+    end
+  end
+
+  # Expectes a file_path string with file_name format /hbxid_mmddyyyy_invoice_r.pdf
+  # Returns Organization
+  def self.get_organization(file_path)
+    file_name = get_file_name(file_path)
+    hbx_id= file_name.split("_")[0]
+    Organization.where(hbx_id: hbx_id).first
+  end
+
+  # Expectes a file_path string with file_name format /hbxid_mmddyyyy_invoice_r.pdf
+  # Returns Date
+  def self.get_invoice_date(file_path)
+    file_name = get_file_name(file_path)
+    date_string= file_name.split("_")[1]
+    Date.strptime(date_string, "%m%d%Y")
+  end
+
+  def self.get_file_name(file_path)
+    File.basename(file_path)
+  end
+
+  def self.invoice_exist?(invoice_date,org)
+    if org.documents.where("date" => invoice_date).count > 0
+      puts "Invoice already exists for the Org"
+      return true
+    end
   end
 
   def office_location_kinds
