--- spec/factories/invoices.rb
@@ -1,5 +1,0 @@
-FactoryGirl.define do
-  factory :invoice do
-    
-  end
-end
